Create AbstractBeaconStateSchema to hold common logic
